Move protection infrastructure to core
The protection infrastructure is useful for other modules like pump drivers,
so it should be in core.
