+++ app/src/main/res/values/strings.xml
@@ -1383,7 +1383,6 @@
     <string name="authorizationfailed">Authorization failed</string>
     <string name="overview_show_absinsulin">Absolute insulin</string>
     <string name="master_password_summary">Master password is used for backup encryption and to override security in application. Remember it or store on a safe place.</string>
-    <string name="passwords_dont_match">Passwords don\'t match</string>
     <string name="current_master_password">Current master password</string>
     <string name="statuslights">Status lights</string>
     <string name="statuslights_copy_ns">Copy settings from NS</string>
