+++ core/build.gradle
@@ -44,6 +44,7 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation "androidx.preference:preference-ktx:1.1.1"
+    implementation 'androidx.biometric:biometric:1.0.1'
     implementation "androidx.activity:activity:${activityVersion}"
     implementation "androidx.activity:activity-ktx:${activityVersion}"
     implementation 'com.google.android.material:material:1.1.0'
