+++ core/src/main/res/drawable/ic_header_key.xml
