+++ core/src/main/res/layout/passwordprompt.xml
