+++ core/src/main/res/values-af-rZA/protection.xml
