+++ core/src/main/res/values-ar-rSA/protection.xml
