+++ core/src/main/res/values-bg-rBG/protection.xml
