+++ core/src/main/res/values-cs-rCZ/protection.xml
