+++ core/src/main/res/values-cy-rGB/protection.xml
