+++ core/src/main/res/values-de-rDE/protection.xml
