+++ core/src/main/res/values-el-rGR/protection.xml
