+++ core/src/main/res/values-es-rES/protection.xml
