+++ core/src/main/res/values-fr-rFR/protection.xml
