+++ core/src/main/res/values-ga-rIE/protection.xml
