+++ core/src/main/res/values-hr-rHR/protection.xml
