+++ core/src/main/res/values-iw-rIL/protection.xml
