+++ core/src/main/res/values-ja-rJP/protection.xml
