+++ core/src/main/res/values-ko-rKR/protection.xml
