+++ core/src/main/res/values-lt-rLT/protection.xml
