+++ core/src/main/res/values-nl-rNL/protection.xml
