+++ core/src/main/res/values-pl-rPL/protection.xml
