+++ core/src/main/res/values-pt-rBR/protection.xml
