+++ core/src/main/res/values-pt-rPT/protection.xml
