+++ core/src/main/res/values-ru-rRU/protection.xml
