+++ core/src/main/res/values-sk-rSK/protection.xml
