+++ core/src/main/res/values-sl-rSI/protection.xml
