+++ core/src/main/res/values-sv-rSE/protection.xml
