+++ core/src/main/res/values-tr-rTR/protection.xml
