+++ core/src/main/res/values-zh-rCN/protection.xml
