+++ core/src/main/res/values/protection.xml
