+++ core/src/main/res/values/strings.xml
@@ -127,6 +127,10 @@
     <string name="shortminute">m</string>
     <string name="shortday">d</string>
 
+    <!--    Protection-->
+    <string name="wrongpassword">Wrong password</string>
+    <string name="passwords_dont_match">Passwords don\'t match</string>
+
     <!--    Profile-->
     <string name="basalprofilenotaligned">Basal values not aligned to hours: %1$s</string>
     <string name="minimalbasalvaluereplaced">Basal value replaced by minimal supported value: %1$s</string>
